Move Connection and Curosr implementation in different modules
