+++ test/unit/exasol/test_dbapi.py
@@ -6,12 +6,14 @@ import importlib
 
 import pytest
 
+from exasol.driver.websocket._connection import _requires_connection
+from exasol.driver.websocket._cursor import (
+    MetaData,
+    _from_pyexasol,
+)
 from exasol.driver.websocket.dbapi2 import (
     Error,
-    MetaData,
     Types,
-    _from_pyexasol,
-    _requires_connection,
 )
